Re #720 Fixed warnings with MinGW Gcc 4.7.1
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@5043 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: b286fc821a81a4aeb3b7e8a1ce9b6ca597fcc3b2
Former-commit-id: c4b48207f587450ed196d7a801fc3e85d9640cae

+++ src/libs/tgf/file.cpp
@@ -110,7 +110,7 @@ bool GfFileCopy(const char* pszSrcName, const char* pszTgtName)
 	// TODO: Use GfFileGetDirName
 	strncpy(buf, pszTgtName, strlen(pszTgtName)+1);
 #ifdef WIN32
-	for (int i = 0; i < maxBufSize && buf[i] != '\0'; i++)
+	for (size_t i = 0; i < maxBufSize && buf[i] != '\0'; i++)
 		if (buf[i] == '\\')
 			buf[i] = '/';
 #endif
